New Branch, Shop Prototype
Making the smallest example of a shop It has Grid of Shop Items that have a title  price and image, clicking buttons is supposed titles buys them then outputs at the end to the terminal. Images are basically for stubs and made a simple implementation of a shop with item bought and price, and buy, new, print statement, functions. It is still incomplete.